feat: improve stemqueue zk-vote process
