+++ components/StemQueue/StemQueue.styles.ts
@@ -1,22 +1,11 @@
 const styles = {
-	actionBtn: {
-		borderWidth: '3px',
-		borderColor: '#111',
-		borderRadius: '5px',
-		fontWeight: 800,
-		my: 4,
-		mr: 2,
-		'&:hover': {
-			borderWidth: '3px',
-		},
-	},
 	noStemsMsg: {
 		my: 4,
 		textAlign: 'center',
 	},
-	stemWrapper: {
-		my: 2,
-		borderTop: '3px solid #000',
+	btn: {
+		width: '56px',
+		fontSize: '10px',
 	},
 	loadingIcon: {
 		my: 0.4,
